# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
  github_com_teleskop_notification-system_pkg_models.Channel:
    enum:
    - sms
    - email
    - push
    type: string
    x-enum-varnames:
    - ChannelSMS
    - ChannelEmail
    - ChannelPush
  github_com_teleskop_notification-system_pkg_models.Notification:
    properties:
      batch_id:
        type: string
      channel:
        $ref: '#/definitions/github_com_teleskop_notification-system_pkg_models.Channel'
      content:
        type: string
      correlation_id:
        type: string
      created_at:
        type: string
      error_message:
        type: string
      id:
        type: string
      idempotency_key:
        type: string
      priority:
        $ref: '#/definitions/github_com_teleskop_notification-system_pkg_models.Priority'
      provider_message_id:
        type: string
      recipient:
        type: string
      retry_count:
        type: integer
      scheduled_at:
        type: string
      sent_at:
        type: string
      status:
        $ref: '#/definitions/github_com_teleskop_notification-system_pkg_models.Status'
      template_vars:
        type: string
      updated_at:
        type: string
    type: object
  github_com_teleskop_notification-system_pkg_models.Priority:
    enum:
    - high
    - normal
    - low
    type: string
    x-enum-varnames:
    - PriorityHigh
    - PriorityNormal
    - PriorityLow
  github_com_teleskop_notification-system_pkg_models.Status:
    enum:
    - pending
    - processing
    - sent
    - failed
    - cancelled
    type: string
    x-enum-varnames:
    - StatusPending
    - StatusProcessing
    - StatusSent
    - StatusFailed
    - StatusCancelled
  services_api_handlers.BatchCreateRequest:
    properties:
      notifications:
        items:
          $ref: '#/definitions/services_api_handlers.CreateNotificationRequest'
        type: array
    type: object
  services_api_handlers.BatchCreateResponse:
    properties:
      batch_id:
        type: string
      notification_ids:
        items:
          type: string
        type: array
    type: object
  services_api_handlers.CreateNotificationRequest:
    properties:
      channel:
        type: string
      content:
        type: string
      priority:
        type: string
      recipient:
        type: string
      scheduled_at:
        type: string
      template_vars:
        additionalProperties:
          type: string
        type: object
    type: object
  services_api_handlers.DeliveryMetrics:
    properties:
      avg_latency_ms:
        type: number
      success_rate:
        type: number
      total_failed:
        type: integer
      total_sent:
        type: integer
    type: object
  services_api_handlers.ErrorResponse:
    properties:
      details:
        type: string
      error:
        type: string
    type: object
  services_api_handlers.HealthResponse:
    properties:
      checks:
        additionalProperties:
          type: string
        type: object
      status:
        type: string
    type: object
  services_api_handlers.MetricsResponse:
    properties:
      delivery:
        $ref: '#/definitions/services_api_handlers.DeliveryMetrics'
      queues:
        additionalProperties:
          $ref: '#/definitions/services_api_handlers.QueueMetrics'
        type: object
    type: object
  services_api_handlers.QueueMetrics:
    properties:
      active:
        type: integer
      failed:
        type: integer
      pending:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: Event-driven notification system for SMS, Email, and Push channels
  title: Notification System API
  version: "1.0"
paths:
  /health:
    get:
      description: Returns the health status of the API and its dependencies
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services_api_handlers.HealthResponse'
      summary: Health check
      tags:
      - health
  /metrics:
    get:
      description: Returns delivery metrics and queue statistics
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services_api_handlers.MetricsResponse'
      summary: Get system metrics
      tags:
      - metrics
  /notifications:
    get:
      description: Returns a cursor-paginated list of notifications with optional
        filters
      parameters:
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Filter by channel
        in: query
        name: channel
        type: string
      - description: Filter by created_at >= from (RFC3339)
        in: query
        name: from
        type: string
      - description: Filter by created_at <= to (RFC3339)
        in: query
        name: to
        type: string
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      - description: Number of results per page
        in: query
        name: page_size
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: List notifications
      tags:
      - notifications
    post:
      consumes:
      - application/json
      description: Create a new notification request
      parameters:
      - description: Idempotency key
        in: header
        name: Idempotency-Key
        type: string
      - description: Notification request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services_api_handlers.CreateNotificationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_teleskop_notification-system_pkg_models.Notification'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/services_api_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/services_api_handlers.ErrorResponse'
      summary: Create a notification
      tags:
      - notifications
  /notifications/{id}:
    get:
      description: Retrieve a single notification by its UUID
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_teleskop_notification-system_pkg_models.Notification'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/services_api_handlers.ErrorResponse'
      summary: Get a notification
      tags:
      - notifications
  /notifications/{id}/cancel:
    patch:
      description: Cancel a notification only if it is in pending status
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_teleskop_notification-system_pkg_models.Notification'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/services_api_handlers.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/services_api_handlers.ErrorResponse'
      summary: Cancel a notification
      tags:
      - notifications
  /notifications/batch:
    post:
      consumes:
      - application/json
      description: Create multiple notifications in a single request
      parameters:
      - description: Idempotency key
        in: header
        name: Idempotency-Key
        type: string
      - description: Batch notification request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services_api_handlers.BatchCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/services_api_handlers.BatchCreateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/services_api_handlers.ErrorResponse'
      summary: Create a batch of notifications
      tags:
      - notifications
  /notifications/batch/{batchId}:
    get:
      description: Retrieve all notifications belonging to a batch
      parameters:
      - description: Batch ID
        in: path
        name: batchId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/services_api_handlers.ErrorResponse'
      summary: Get notifications by batch ID
      tags:
      - notifications
swagger: "2.0"